Implement senior unlock sensibly and make it work for age-encrypted identities
Before it just asked the passphrase once, assumed it was correct, and
saved it in the agent. This never works for age-encrypted identities and
for encrypted ssh keys it only works if you enter the correct
passphrase.
This is because for age-encrypted identities senior-agent does not store
the passphrase, but the whole identity for performance reasons. When
senior gets an identity from senior-agent it immediately assumes it is
correct and uses that.