Check for merge conflicts before reencryption
`senior reencrypt` and `senior add-recipient` both reencrypt the entire
store. If the remote branch is >= 1 commit ahead of the local branch
then a reencrypt of the entire store almost guarantees merge conflicts
on the next pull.
This is why senior now does a `git fetch` beforehand. If the remote is
ahead of the local branch, or if the fetch did not work, then a warning
is shown and the user is asked if they want to continue anyway.